fix tests across the board
