+++ packages/core/src/test/sequence-validation.test.tsx
@@ -2,46 +2,11 @@
  * @vitest-environment jsdom
  */
 import {render} from '@testing-library/react';
-import React from 'react';
 import {describe, expect, test} from 'vitest';
 import {CanUseRemotionHooksProvider} from '../CanUseRemotionHooks';
-import type {CompositionManagerContext} from '../CompositionManager';
-import {CompositionManager} from '../CompositionManager';
 import {Sequence} from '../Sequence';
 import {expectToThrow} from './expect-to-throw';
-
-const Comp: React.FC = () => null;
-
-const context: CompositionManagerContext = {
-	assets: [],
-	compositions: [
-		{
-			id: 'my-comp',
-			durationInFrames: 100,
-			// @ts-expect-error
-			component: Comp,
-			defaultProps: {},
-			folderName: null,
-			fps: 30,
-			height: 1080,
-			width: 1080,
-			parentFolderName: null,
-			nonce: 0,
-		},
-	],
-	currentComposition: 'my-comp',
-	folders: [],
-	registerAsset: () => undefined,
-	registerComposition: () => undefined,
-	registerFolder: () => undefined,
-	registerSequence: () => undefined,
-	sequences: [],
-	setCurrentComposition: () => undefined,
-	unregisterAsset: () => undefined,
-	unregisterComposition: () => undefined,
-	unregisterFolder: () => undefined,
-	unregisterSequence: () => undefined,
-};
+import {WrapSequenceContext} from './wrap-sequence-context';
 
 describe('Composition-validation render should throw with invalid props', () => {
 	describe('Throw with invalid duration props', () => {
@@ -102,11 +67,11 @@ describe('Composition-validation render should NOT throw with valid props', () =
 		expect(() =>
 			render(
 				<CanUseRemotionHooksProvider>
-					<CompositionManager.Provider value={context}>
+					<WrapSequenceContext>
 						<Sequence durationInFrames={100} from={0}>
 							{null}
 						</Sequence>
-					</CompositionManager.Provider>
+					</WrapSequenceContext>
 				</CanUseRemotionHooksProvider>
 			)
 		).not.toThrow();
@@ -115,11 +80,11 @@ describe('Composition-validation render should NOT throw with valid props', () =
 		expect(() =>
 			render(
 				<CanUseRemotionHooksProvider>
-					<CompositionManager.Provider value={context}>
+					<WrapSequenceContext>
 						<Sequence durationInFrames={100} from={0}>
 							{undefined}
 						</Sequence>
-					</CompositionManager.Provider>
+					</WrapSequenceContext>
 				</CanUseRemotionHooksProvider>
 			)
 		).not.toThrow();
